Opening ceremony Bath's 2023 Quantum Bootcamp
Hackathon weekend event open to University of Bath Science students with the objective to promote Quantum technologies and foster interest and experience in Quantum Software. This presentation was used for the opening ceremony which I presented. It contains relevant event based information as well as introductory slides for certain quantum hardware technologies, event sponsors and speaker.